+++ patches/Accessibility/strings_de.cpp
@@ -355,6 +355,7 @@ const char* Get(Id id) {
         case Id::AxisNorthSouth:             return "Nord-S\xFC""d";   // Nord-Süd
         case Id::AxisEastWest:               return "Ost-West";
         case Id::AreaNoun:                   return "Bereich";
+        case Id::AreaNounLarge:              return "Gro\xDF""er Bereich";  // Großer Bereich
         case Id::FmtAreaAxisExits:           return "%s %s. Ausg\xE4""nge: %s";
         case Id::FmtAreaExits:               return "%s. Ausg\xE4""nge: %s";
         case Id::FmtAreaAxisOnly:            return "%s %s";